requirements.txt
django
djangorestframework
gunicorn
whitenoise
django-cors-headers
django-extensions
djangorestframework-simplejwt[crypto]
